reworked run_command as module
